feat: extract hooks into hudozka-hooks package
