+++ apps/hudozka/package.json
@@ -19,27 +19,24 @@
     "author": "Roman Timashev",
     "license": "ISC",
     "dependencies": {
+        "@hudozka/hooks": "*",
+        "@hudozka/text": "*",
         "@hudozka/utils": "*",
         "@radix-ui/react-navigation-menu": "^1.2.0",
         "@radix-ui/react-tooltip": "^1.2.0",
         "classnames": "^2.5.0",
         "dangerously-set-html-content": "^1.1.0",
-        "markdown-it": "^14.1.1",
         "next": "^16.2.0",
         "pocketbase": "^0.26.0",
         "react": "^19.2.0",
         "react-dom": "^19.2.0",
         "react-icons": "^5.5.0",
         "sharp": "^0.34.0",
-        "typograf": "^7.6.0",
         "valtio": "^2.3.0"
     },
     "overrides": {
         "dangerously-set-html-content": {
             "react": "$react"
         }
-    },
-    "devDependencies": {
-        "@types/markdown-it": "^14.1.0"
     }
 }
