+++ apps/hudozka/src/components/App/index.tsx
@@ -8,7 +8,7 @@ import { Content } from "../Content"
 import { Breadcrumbs } from "../Breadcrumbs"
 import type { BreadcrumbPart } from "@/types"
 import { usePathname } from "next/navigation"
-import { useMobile } from "@/hooks/useMobile"
+import { useMobile } from "@hudozka/hooks"
 import { Navigation } from "@/components/Navigation"
 import { Box } from "@/ui/Box"
 import { AccessibilityButton } from "../AccessibilityButton"
