+++ apps/hudozka/src/components/PageGrid/index.tsx
@@ -6,7 +6,7 @@ import type { CardLayout } from "../Card"
 import { Date } from "./Date"
 import { Spacer } from "../Spacer"
 import { Box } from "@/ui/Box"
-import { useMobile } from "@/hooks/useMobile"
+import { useMobile } from "@hudozka/hooks"
 
 function itemColumn(item: PageCardDto): string {
     return item.featured ? "span 2" : "auto"
