+++ apps/hudozka/src/hooks/useDarkTheme.ts
@@ -1,4 +1,4 @@
-import { useMediaQuery } from "@/hooks/useMediaQuery"
+import { useMediaQuery } from "@hudozka/hooks"
 
 export function useDarkTheme(): boolean {
     return useMediaQuery("(prefers-color-scheme: dark)")
