+++ docs/plans/20260424-monorepo.md
@@ -971,7 +971,7 @@ Goal: the remaining shared packages exist and are consumed by the web app; all t
 - Move: `apps/hudozka/src/lib/text.test.ts` → `packages/text/src/text.test.ts`
 - Create: `packages/text/src/index.ts`
 
-- [ ] **Step 1:** Create `packages/text/package.json`:
+- [x] **Step 1:** Create `packages/text/package.json`:
 
 ```json
 {
@@ -995,7 +995,7 @@ Goal: the remaining shared packages exist and are consumed by the web app; all t
 }
 ```
 
-- [ ] **Step 2:** Create `packages/text/tsconfig.json`:
+- [x] **Step 2:** Create `packages/text/tsconfig.json`:
 
 ```json
 {
@@ -1004,7 +1004,7 @@ Goal: the remaining shared packages exist and are consumed by the web app; all t
 }
 ```
 
-- [ ] **Step 3:** Move the files.
+- [x] **Step 3:** Move the files.
 
 ```bash
 cd /Users/tmshv/Workspace/__github_tmshv/hudozka
@@ -1013,13 +1013,13 @@ git mv apps/hudozka/src/lib/text.ts packages/text/src/text.ts
 git mv apps/hudozka/src/lib/text.test.ts packages/text/src/text.test.ts
 ```
 
-- [ ] **Step 4:** Create `packages/text/src/index.ts`:
+- [x] **Step 4:** Create `packages/text/src/index.ts`:
 
 ```ts
 export * from "./text"
 ```
 
-- [ ] **Step 5:** Codemod imports.
+- [x] **Step 5:** Codemod imports.
 
 ```bash
 rg -l '@/lib/text' apps/hudozka/src \
@@ -1028,7 +1028,7 @@ rg '@/lib/text' apps/hudozka
 ```
 Expected: no matches in the second command.
 
-- [ ] **Step 6:** Update `apps/hudozka/package.json`:
+- [x] **Step 6:** Update `apps/hudozka/package.json`:
 - Add: `"@hudozka/text": "*",`
 - Remove: `"markdown-it": "^14.1.1",` and `"typograf": "^7.6.0",` from `dependencies`
 - Remove: `"@types/markdown-it": "^14.1.0"` from `devDependencies` (now a dev dep of the package)
@@ -1037,7 +1037,7 @@ Expected: no matches in the second command.
 npm install
 ```
 
-- [ ] **Step 7:** Verify.
+- [x] **Step 7:** Verify.
 
 ```bash
 npm test -w @hudozka/text
@@ -1046,12 +1046,14 @@ npm run build
 ```
 Expected: all pass.
 
-- [ ] **Step 8:** Commit.
+- [x] **Step 8:** Commit.
 
 ```bash
 git add packages/text apps/hudozka package.json package-lock.json
 git commit -m "Extract markdown/typograf rendering into @hudozka/text"
 ```
+
+Notes: `@hudozka/text` tests pass (4/4, 1 file). `apps/hudozka` tests pass (55/55, 11 files). `npm run build` succeeds — the "Failed to fetch" CMS errors during build are the documented expected behaviour for local builds (CMS unreachable). Only one source file in the app imported `@/lib/text` (`apps/hudozka/src/remote/factory.ts`); rewritten to `@hudozka/text`. No other references to `markdown-it`, `typograf`, or `@/lib/text` remain in `apps/hudozka/src`. `markdown-it` and `typograf` moved to `packages/text/package.json` as runtime deps; `@types/markdown-it` moved there as a devDependency. All three dropped from `apps/hudozka/package.json`.
 
 ### Task 5.2: Scaffold and populate `@hudozka/hooks`
 
@@ -1068,7 +1070,7 @@ git commit -m "Extract markdown/typograf rendering into @hudozka/text"
   - `useReducedMotion.ts` (no test exists)
 - Create: `packages/hooks/src/index.ts`
 
-- [ ] **Step 1:** Create `packages/hooks/package.json`:
+- [x] **Step 1:** Create `packages/hooks/package.json`:
 
 ```json
 {
@@ -1088,7 +1090,7 @@ git commit -m "Extract markdown/typograf rendering into @hudozka/text"
 }
 ```
 
-- [ ] **Step 2:** Create `packages/hooks/tsconfig.json`:
+- [x] **Step 2:** Create `packages/hooks/tsconfig.json`:
 
 ```json
 {
@@ -1097,7 +1099,7 @@ git commit -m "Extract markdown/typograf rendering into @hudozka/text"
 }
 ```
 
-- [ ] **Step 3:** Create `packages/hooks/vitest.config.ts`:
+- [x] **Step 3:** Create `packages/hooks/vitest.config.ts`:
 
 ```ts
 import { defineConfig } from "vitest/config"
@@ -1111,13 +1113,13 @@ export default defineConfig({
 })
 ```
 
-- [ ] **Step 4:** Create `packages/hooks/vitest.setup.ts`:
+- [x] **Step 4:** Create `packages/hooks/vitest.setup.ts`:
 
 ```ts
 import "@testing-library/jest-dom/vitest"
 ```
 
-- [ ] **Step 5:** Move the files.
+- [x] **Step 5:** Move the files.
 
 ```bash
 cd /Users/tmshv/Workspace/__github_tmshv/hudozka
@@ -1132,14 +1134,14 @@ git mv apps/hudozka/src/hooks/useMobile.ts packages/hooks/src/useMobile.ts
 git mv apps/hudozka/src/hooks/useReducedMotion.ts packages/hooks/src/useReducedMotion.ts
 ```
 
-- [ ] **Step 6:** Confirm `useDarkTheme.ts` and `useAccessibility.ts` remain in `apps/hudozka/src/hooks/`.
+- [x] **Step 6:** Confirm `useDarkTheme.ts` and `useAccessibility.ts` remain in `apps/hudozka/src/hooks/`.
 
 ```bash
 ls apps/hudozka/src/hooks
 ```
 Expected: `useAccessibility.ts`, `useDarkTheme.ts` only.
 
-- [ ] **Step 7:** Create `packages/hooks/src/index.ts`:
+- [x] **Step 7:** Create `packages/hooks/src/index.ts`:
 
 ```ts
 export * from "./useToggle"
@@ -1149,7 +1151,7 @@ export * from "./useMobile"
 export * from "./useReducedMotion"
 ```
 
-- [ ] **Step 8:** Codemod imports in the web app. Only the generic hook paths change; business-coupled hook imports (`useDarkTheme`, `useAccessibility`) stay pointing at `@/hooks/`.
+- [x] **Step 8:** Codemod imports in the web app. Only the generic hook paths change; business-coupled hook imports (`useDarkTheme`, `useAccessibility`) stay pointing at `@/hooks/`.
 
 ```bash
 cd /Users/tmshv/Workspace/__github_tmshv/hudozka
@@ -1159,7 +1161,7 @@ rg '@/hooks/(useToggle|useLockBodyScroll|useMediaQuery|useMobile|useReducedMotio
 ```
 Expected: no matches in the second command.
 
-- [ ] **Step 9:** Collapse any duplicate `import … from "@hudozka/hooks"` lines in affected files.
+- [x] **Step 9:** Collapse any duplicate `import … from "@hudozka/hooks"` lines in affected files.
 
 ```bash
 rg -l '"@hudozka/hooks"' apps/hudozka/src | while read f; do
@@ -1169,13 +1171,13 @@ done
 ```
 For each file listed, merge imports into a single line.
 
-- [ ] **Step 10:** Update `apps/hudozka/package.json` — add `"@hudozka/hooks": "*",` to dependencies.
+- [x] **Step 10:** Update `apps/hudozka/package.json` — add `"@hudozka/hooks": "*",` to dependencies.
 
 ```bash
 npm install
 ```
 
-- [ ] **Step 11:** Verify.
+- [x] **Step 11:** Verify.
 
 ```bash
 npm test -w @hudozka/hooks
@@ -1184,12 +1186,14 @@ npm run build
 ```
 Expected: all pass.
 
-- [ ] **Step 12:** Commit.
+- [x] **Step 12:** Commit.
 
 ```bash
 git add packages/hooks apps/hudozka package.json package-lock.json
 git commit -m "Extract generic React hooks into @hudozka/hooks"
 ```
+
+Notes: `@hudozka/hooks` tests pass (13/13 across 3 files). `apps/hudozka` tests pass (42/42 across 8 files). `npm run build` succeeds — Next compiles, TypeScript passes, the standalone output is produced; the "Failed to fetch" CMS errors during static generation are the documented expected behaviour for local builds (CMS unreachable). Internal imports in `useMobile.ts` and `useReducedMotion.ts` (previously `@/hooks/useMediaQuery`) were rewritten to relative `./useMediaQuery` since the package has no `@/` path alias. `useDarkTheme.ts` (remaining in `apps/hudozka/src/hooks/`) was updated from `@/hooks/useMediaQuery` to `@hudozka/hooks` so the app still resolves its transitive hook dependency. One stale legacy-form `src/hooks/useReducedMotion` import (in `components/Content/index.tsx`) that the plan regex didn't catch was also rewritten to `@hudozka/hooks` for the build to pass. Two `@hudozka/hooks` imports in `ui/Overlay/index.tsx` were merged into a single line. `npm install` symlinked `@hudozka/hooks` at the root `node_modules/@hudozka/` tree (hoisted). Also commits the staged-but-uncommitted Task 5.1 work from the previous iteration (the @hudozka/text scaffold + factory.ts rewrite); see commit message.
 
 ### Task 5.3: Scaffold and populate `@hudozka/ui`
