+++ packages/hooks/package.json
@@ -1,0 +1,15 @@
+{
+    "name": "@hudozka/hooks",
+    "private": true,
+    "version": "0.0.0",
+    "type": "module",
+    "exports": {
+        ".": "./src/index.ts"
+    },
+    "scripts": {
+        "test": "vitest run"
+    },
+    "peerDependencies": {
+        "react": ">=19"
+    }
+}
