+++ packages/hooks/src/useMobile.ts
@@ -1,4 +1,4 @@
-import { useMediaQuery } from "@/hooks/useMediaQuery"
+import { useMediaQuery } from "./useMediaQuery"
 
 export function useMobile(): boolean {
     return useMediaQuery("(max-width: 31.25em)", false)
