+++ packages/hooks/src/useReducedMotion.ts
@@ -1,4 +1,4 @@
-import { useMediaQuery } from "@/hooks/useMediaQuery"
+import { useMediaQuery } from "./useMediaQuery"
 
 export function useReducedMotion(): boolean {
     return useMediaQuery("(prefers-reduced-motion: reduce)")
