+++ packages/hooks/vitest.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from "vitest/config"
+
+export default defineConfig({
+    test: {
+        include: ["src/**/*.test.{ts,tsx}"],
+        environment: "happy-dom",
+        setupFiles: ["./vitest.setup.ts"],
+    },
+})
