+++ packages/text/package.json
@@ -1,0 +1,19 @@
+{
+    "name": "@hudozka/text",
+    "private": true,
+    "version": "0.0.0",
+    "type": "module",
+    "exports": {
+        ".": "./src/index.ts"
+    },
+    "scripts": {
+        "test": "vitest run"
+    },
+    "dependencies": {
+        "markdown-it": "^14.1.0",
+        "typograf": "^7.6.0"
+    },
+    "devDependencies": {
+        "@types/markdown-it": "^14.1.0"
+    }
+}
