+++ packages/text/src/text.ts
